Fixing compilation on linux

+++ Source/Common/fileutil.cpp
@@ -296,7 +296,7 @@ void fsetmode(FILE* f, char type)
 
 void freadOrDie(void* ptr, size_t size, size_t count, FILE* f)
 {
-    size_t limit = max(READ_SIZE_LIMIT / size, 1ull);  // Normalize by size, as fread() expects units, not bytes
+    size_t limit = max(READ_SIZE_LIMIT / size, (size_t)1);  // Normalize by size, as fread() expects units, not bytes
 
     // \\XXX\C$ reads are limited, with some randomness (e.g. 48 MB), on Windows 7 32 bit, so we break this into chunks of some MB. Meh.
     while (count > 0)
